requirements.txt
# Core LangChain and AI libraries
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.6.1

# Web framework and API
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Speech recognition and synthesis
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11

# HTTP requests and data handling
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# Logging and utilities
loguru==0.7.2
pydantic==2.5.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
